int to int 8

+++ 0.0002/queue/prorityQueue.go
@@ -33,7 +33,7 @@ func prorityQueue() {
 		}
 	}
 }
-func EnqueueInProrityQueue() (int, int) {
+func EnqueueInProrityQueue() (int, int8) {
 	fmt.Println("Enter the value to Enqueue in Prority Queue")
 	fmt.Println("In format: value,prority")
 	fmt.Println("Note: Prority shall be from 0 to 4, lower the value, higher the prority")
